Moved logic into Game model
