+++ HausClock/Models/Game.swift
@@ -7,9 +7,49 @@
 //
 
 import Foundation
+import Dollar
 
 enum GameState {
     case Active
     case Paused
     case Finished
+}
+
+class Game {
+    var state = GameState.Paused
+    
+    init() {
+        reset()
+    }
+    
+    var players = [
+        Player(position: .Top),
+        Player(position: .Bottom)
+    ]
+    
+    func reset() {
+        for player in players {
+            player.secondsRemaining = initialTimeInSeconds
+        }
+
+        setPlayerToActive(.Top)
+        state = .Paused
+    }
+    
+    func setPlayerToActive(position: PlayerPosition) {
+        var activePlayer = getPlayerByPosition(position)
+        var inactivePlayer = getPlayerByPosition(position.opposite())
+        
+        activePlayer.state = .Active
+        inactivePlayer.state = .Waiting
+        state = .Active
+    }
+    
+    func getPlayerByPosition(position: PlayerPosition) -> Player {
+        return $.find(players, { $0.position == position } )!
+    }
+    
+    func getActivePlayer() -> Player? {
+        return $.find(players, { $0.state == PlayerState.Active } )!
+    }
 }
